Exploring using execution resources for allocation
The current execution context interface relies on reference-based
execution resources. However, this complicates the construction of
objects that need to store the information.
The current hwlocxx_context interface relies on the resources to store
the pointer to the topology, which means if the object is not available,
the pointer is missing and allocators cannot be created.

Implementation alternative:
* Store the topology pointer not in the resource but in the context,
   however that implies recreating the topology every time a new
   execution context is created, which is not feasible.